Support node comments on generic nodes
